refactor(utils): add getDiffInCalendarHours as ESM
