+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export { default as getDiffInMilliseconds } from './getDiffInMilliseconds.js';
 export { default as getDiffInCalendarSeconds } from './getDiffInCalendarSeconds.js';
 export { default as getDiffInCalendarMinutes } from './getDiffInCalendarMinutes.js';
+export { default as getDiffInCalendarHours } from './getDiffInCalendarHours.js';
